tests passing, changing name to ForwardInfluenceBlanket
